requirements.txt
python-dotenv==1.0.1
langchain==0.3.21
openai==1.66.3
langchain-openai==0.3.8
supabase==2.13.0
fastapi==0.115.11
python-multipart==0.0.20
langchain-community==0.3.20
pydantic-settings==2.8.1
uvicorn=0.34.0
pypdf==5.4.0
langgraph==0.3.18
